+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/model/FapInstalledScreenState.kt
@@ -7,7 +7,7 @@ sealed class FapInstalledScreenState {
     object Loading : FapInstalledScreenState()
 
     data class Loaded(
-        val faps: ImmutableList<FapItemShort>
+        val faps: ImmutableList<Pair<FapItemShort, FapInstalledInternalState>>
     ) : FapInstalledScreenState()
 
     data class LoadedOffline(